+++ IdjGameEngine/include/Minion.h
@@ -4,7 +4,7 @@
 
 class Minion : public Component {
 public:
-    Minion(GameObject& associated, const std::weak_ptr<GameObject>& alienCenter, double arcOffset = 0);
+    Minion(GameObject& associated, const std::weak_ptr<GameObject>& alienCenter, float arcOffset = 0);
 
     void Start() override;
     void Update(float dt) override;
