+++ IdjGameEngine/src/Minion.cpp
@@ -8,12 +8,12 @@
 #include "Constants.h"
 #include "GameObject.h"
 
-Minion::Minion(GameObject& associated, const std::weak_ptr<GameObject>& alienCenter, double arcOffset):
+Minion::Minion(GameObject& associated, const std::weak_ptr<GameObject>& alienCenter, float arcOffset):
     Component(associated),
     alienCenter(alienCenter),
     arc(arcOffset) {
 
-    auto scale = 1 + ((rand() % 5) + 1) / 10.0;
+    auto scale = float(1 + ((rand() % 5) + 1) / 10.0);
 
     auto sprite = std::make_shared<Sprite>(associated, "assets/img/minion.png");
     sprite->SetScale(scale, scale);
@@ -54,7 +54,7 @@ void Minion::Shoot(Vec2 target) const {
     auto bulletGO = std::make_shared<GameObject>();
 
     bulletGO->AddComponent(std::make_shared<Bullet>(*bulletGO, angle, Constants::Minion::BULLET_SPEED, Constants::Minion::BULLET_DAMAGE,
-                                                    Constants::Minion::BULLET_MAX_DISTANCE, "assets/img/minionbullet2.png", 3, 1, true));
+                                                    Constants::Minion::BULLET_MAX_DISTANCE, "assets/img/minionbullet2.png", 3, 1.0f, true));
 
     bulletGO->box.SetCenter(position.x, position.y);
     bulletGO->angleDeg = Util::RadToDeg(angle);
